backendenergy/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
astroid==2.6.5; python_version ~= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==21.3.1; python_version >= '3.7'
automat==20.2.0
cffi==1.14.6
channels==3.0.4; python_version >= '3.6'
constantly==15.1.0
cryptography==3.4.7; python_version >= '3.6'
daphne==3.0.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-guardian==2.4.0
django==3.2.5
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
et-xmlfile==1.1.0; python_version >= '3.6'
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
isort==5.9.2; python_version < '4.0' and python_full_version >= '3.6.1'
lazy-object-proxy==1.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
markdown==3.3.4
mccabe==0.6.1
openpyxl==3.0.7
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.9.0
pylint-django==2.4.4
pylint-plugin-utils==0.6
pylint==2.9.6
pyopenssl==20.0.1
pytz==2021.1
pyyaml==5.4.1
service-identity==21.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
twisted[tls]==21.7.0; python_full_version >= '3.6.7'
txaio==21.2.1; python_version >= '3.6'
typing-extensions==3.10.0.0
uritemplate==3.0.1
wrapt==1.12.1
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
